Feat/game service

+++ src/main/java/at/fhtw/mtcg_app/service/trading/TradeController.java
@@ -31,6 +31,7 @@ public class TradeController extends Controller {
         String username = request.getTokenUser();
 
         if(username.equals("")){
+            newUnit.close();
             return new Response(
                     HttpStatus.UNAUTHORIZED,
                     ContentType.PLAIN_TEXT,
@@ -40,6 +41,7 @@ public class TradeController extends Controller {
         try{
             boolean userExists=this.userRepo.checkUserExists(username, newUnit);
             if(!userExists){
+                newUnit.close();
                 return new Response(
                         HttpStatus.NOT_FOUND,
                         ContentType.PLAIN_TEXT,
@@ -50,6 +52,7 @@ public class TradeController extends Controller {
             String dealID=request.getPathParts().get(1);
 
             if(!this.tradingRepo.tradingDealAlreadyExists(dealID, newUnit)){
+                newUnit.close();
                 return new Response(
                         HttpStatus.NOT_FOUND,
                         ContentType.PLAIN_TEXT,
@@ -67,6 +70,7 @@ public class TradeController extends Controller {
 
 
             if(!this.cardsRepo.checkIfCardsBelongToUser(username, tradeCard, newUnit)){
+                newUnit.close();
                 return new Response(
                         HttpStatus.FORBIDDEN,
                         ContentType.PLAIN_TEXT,
@@ -104,6 +108,7 @@ public class TradeController extends Controller {
         String username = request.getTokenUser();
 
         if(username.equals("")){
+            newUnit.close();
             return new Response(
                     HttpStatus.UNAUTHORIZED,
                     ContentType.PLAIN_TEXT,
@@ -113,6 +118,7 @@ public class TradeController extends Controller {
         try {
             boolean userExists = this.userRepo.checkUserExists(username, newUnit);
             if (!userExists) {
+                newUnit.close();
                 return new Response(
                         HttpStatus.NOT_FOUND,
                         ContentType.PLAIN_TEXT,
@@ -124,6 +130,7 @@ public class TradeController extends Controller {
 
             //if deal does not exist
             if (!this.tradingRepo.tradingDealAlreadyExists(dealID, newUnit)) {
+                newUnit.close();
                 return new Response(
                         HttpStatus.NOT_FOUND,
                         ContentType.PLAIN_TEXT,
@@ -141,6 +148,7 @@ public class TradeController extends Controller {
                 String newOwnerOfferCard=this.tradingRepo.getCreatorFromDeal(dealID, newUnit);
                 String offerCard=request.getBody();
                 String dealCard=this.tradingRepo.getCardToTrade(dealID, newUnit);
+
 
                 this.tradingRepo.executeTrade(dealID, newOwnerDealCard, newOwnerOfferCard, dealCard, offerCard, newUnit);
 
@@ -181,6 +189,7 @@ public class TradeController extends Controller {
         String username = request.getTokenUser();
 
         if(username.equals("")){
+            newUnit.close();
             return new Response(
                     HttpStatus.UNAUTHORIZED,
                     ContentType.PLAIN_TEXT,
@@ -191,6 +200,7 @@ public class TradeController extends Controller {
         try{
             boolean userExists=this.userRepo.checkUserExists(username, newUnit);
             if(!userExists){
+                newUnit.close();
                 return new Response(
                         HttpStatus.NOT_FOUND,
                         ContentType.PLAIN_TEXT,
@@ -200,6 +210,7 @@ public class TradeController extends Controller {
             Trade tradeDeal = this.getObjectMapper().readValue(request.getBody(), Trade.class);
 
             if(this.tradingRepo.tradingDealAlreadyExists(tradeDeal.getId(), newUnit)){
+                newUnit.close();
                 return new Response(
                         HttpStatus.CONFLICT,
                         ContentType.PLAIN_TEXT,
@@ -212,6 +223,7 @@ public class TradeController extends Controller {
 
 
             if((!this.cardsRepo.checkIfCardsBelongToUser(username, tradeCard, newUnit)) || this.tradingRepo.checkCardIsLockedInDeck(tradeDeal.getCardToTrade(), newUnit)){
+                newUnit.close();
                 return new Response(
                         HttpStatus.FORBIDDEN,
                         ContentType.PLAIN_TEXT,
@@ -249,6 +261,7 @@ public class TradeController extends Controller {
         String username = request.getTokenUser();
 
         if(username.equals("")){
+            newUnit.close();
             return new Response(
                     HttpStatus.UNAUTHORIZED,
                     ContentType.PLAIN_TEXT,
@@ -259,6 +272,7 @@ public class TradeController extends Controller {
         try{
             boolean userExists=this.userRepo.checkUserExists(username, newUnit);
             if(!userExists){
+                newUnit.close();
                 return new Response(
                         HttpStatus.NOT_FOUND,
                         ContentType.PLAIN_TEXT,
@@ -269,6 +283,7 @@ public class TradeController extends Controller {
             List<Trade> tradeDeals = this.tradingRepo.getTradingDeals(newUnit);
 
             if(tradeDeals.isEmpty()){
+                newUnit.close();
                 return new Response(
                         HttpStatus.NO_CONTENT,
                         ContentType.PLAIN_TEXT,
@@ -277,6 +292,7 @@ public class TradeController extends Controller {
             }
             String cardsJSON = this.getObjectMapper().writeValueAsString(tradeDeals);
 
+            newUnit.close();
             return new Response(
                     HttpStatus.OK,
                     ContentType.JSON,
@@ -288,6 +304,7 @@ public class TradeController extends Controller {
             e.printStackTrace();
         }
 
+        newUnit.close();
         return new Response(
                 HttpStatus.INTERNAL_SERVER_ERROR,
                 ContentType.JSON,
@@ -296,6 +313,7 @@ public class TradeController extends Controller {
     }
 
     private boolean checkIfTradeIsPossible(String dealID, String cardFromOfferID, String userFromOffer, UnitOfWork newUnit) throws Exception {
+
         try{
             String cardFromDealID=this.tradingRepo.getCardToTrade(dealID, newUnit);
 
@@ -305,10 +323,12 @@ public class TradeController extends Controller {
                 return false;
             }
 
+
             if(this.tradingRepo.checkCardIsLockedInDeck(cardFromDealID, newUnit)){
                 this.tradingRepo.deleteTradingDeal(dealID, newUnit);
                 return false;
             }
+
 
 
             List<String> cardFromDealList =new ArrayList<>();
@@ -318,16 +338,21 @@ public class TradeController extends Controller {
 
             String creatorOfDeal=this.tradingRepo.getCreatorFromDeal(dealID, newUnit);
 
+            if(creatorOfDeal.equals(userFromOffer)){
+                return false;
+            }
 
 
             if(!this.cardsRepo.checkIfCardsBelongToUser(userFromOffer, cardFromOfferList, newUnit)){
                 return false;
             }
 
+
             if(!this.cardsRepo.checkIfCardsBelongToUser(creatorOfDeal, cardFromDealList, newUnit)){
                 this.tradingRepo.deleteTradingDeal(dealID, newUnit);
                 return false;
             }
+
 
 
             Trade trade=this.tradingRepo.getTrade(dealID, newUnit);
@@ -338,9 +363,11 @@ public class TradeController extends Controller {
             if(requiredDamage>actualDamage || !requiredType.equals(actualType)){
                 return false;
             }
+
         }catch(Exception e){
             throw new Exception("Could not check if trade is possible");
         }
+
         return true;
     }
 }
